Fixes after conversation in Discord.
We won't make `expectEqualSlices` more generic, because that would
force us to go through every case for each possible type that can be
passed, and check that they are allowed. What makes this more difficult
is that it would need to implement the same implicit type casting that
the type checker does (from string pointer to array of character for
example). First, that amount of work isn't really warranted here, and
second, it's going to be quite a bit of maintenance to make sure that
the rules of implicit type casting continue to be mirrored correctly
in that new function.
Reusing `expectEqualSlices`, and calling it, doesn't really make the
code simpler. It makes the code larger because we have to handle the
different cases where one param is a pointer and the other is an array
and cast them accordingly so that `expectEqualSlices` can do its thing.
But "its thing" is super simple, so it's just more readable to copy that
logic.

So we'll just have this new and different function that looks at
sentinels.