stage2: decl equality so &a == &a in comptime
In stage2, decl ref equality was `decl == decl2` (pointer comparison).
But it is possible for two identical decls to have two separate pointers
via the `ref` ZIR. Each `ref` ZIR creates a new anonymous decl.
Therefore, `&a == &a` would return false in comptime.

This modifies the comparison so that if the pointers are not equivalent
(they are in the fast path), then we check if they were created in the
same scope from the same AST node. If so, we assume they're equivalent.