Mutex for cache access, add addDepFilePostContents
Throwing things at the wall to see what sticks - I don't think all of
these changes are actually needed, but it gives us a baseline to work
from as this did work for me locally. Will check these changes in and
push them - and if they pass the automated tests on the server we can
then start peeling things back until we find the subset of changes
needed to get things working.